fixed all formatting in files, added comments, tried to make leaderboard work, personal and all tasks are working, score changes when ccheckbox is clicked on
