+++ app/(tabs)/groups.tsx
@@ -1,64 +1,66 @@
 import {Text, TextInput, View, StyleSheet, TouchableOpacity, Alert,FlatList } from 'react-native';
-
 import CanDoScrollView from '@/components/CanDoScrollView';
 import { AddTaskPane } from '@/components/AddTaskPane';
 import { TabBarIcon } from '@/components/navigation/TabBarIcon';
 import { Colors } from '@/constants/Colors';
 import {useState, useEffect} from 'react'
-
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import {createStaticNavigation,useNavigation, useRoute} from '@react-navigation/native';
 import { collection, addDoc, getFirestore, app, onSnapshot, query, where } from "firebase/firestore";
 
+//database connection
 const db = getFirestore(app)
 
 export default function GroupsScreen() {
-  const navigation = useNavigation();
-  const [groups, setGroups] = useState([])
-  const [user, setUser] = useState([])
-  const route = useRoute()
+    const navigation = useNavigation();
+    const route = useRoute()
+    //state variables
+    const [groups, setGroups] = useState([])
+    const [user, setUser] = useState([])
 
-  useEffect(() => {
-      const getGroups = collection(db, 'Groups')
+    //use effect to get groups from the database
+    useEffect(() => {
+        const getGroups = collection(db, 'Groups')
 
-      const subscribe = onSnapshot(getGroups, (querySnapshot) => {
-          const addGroups = []
-          querySnapshot.forEach((list) => {
-              addGroups.push({...list.data(), id: list.id})
-          })
-          setGroups(addGroups)
-          console.log(addGroups)
-      });
-      return () => subscribe();
+        const subscribe = onSnapshot(getGroups, (querySnapshot) => {
+            const addGroups = []
+            querySnapshot.forEach((list) => {
+                addGroups.push({...list.data(), id: list.id})
+            })
+            setGroups(addGroups)
+            console.log(addGroups)
+        });
+        return () => subscribe();
     }, []);
 
-  return (
-    <CanDoScrollView>
-        <View style={styles.buttonContainer}>
-        {groups.map((group) => (
-
-               <TouchableOpacity style={[styles.button, {backgroundColor: group.color}]} onPress={() => navigation.navigate('groupDetail', {groupVal : group.id, groupName: group.groupName, groupUser: group.users,
+    return (
+        <CanDoScrollView>
+            <View style={styles.buttonContainer}>
+                {/*gets the all groups from the database adn displays it*/}
+                {groups.map((group) => (
+                    <TouchableOpacity style={[styles.button, {backgroundColor: group.color}]} onPress={() => navigation.navigate('groupDetail', {groupVal : group.id, groupName: group.groupName, groupUser: group.users,
                                                                                                                                             groupDescription: group.description, groupBool:group.passwordBool, groupPassword: group.password })} >
-                    <Text style={styles.titleText}>{group.groupName}</Text>
-                    {group.users ?(
-                    group.users.map((user, index) => (
-                        <Text style={styles.buttonText} key={index}>
-                        {user}
-                        </Text>
-                    ))
-                    ):(
-                    <Text style={styles.buttonText}> No Users </Text>
-                    )}
-               </TouchableOpacity>
-        ))}
+                        <Text key = {user.id} style={styles.titleText}>{group.groupName}</Text>
+                        {group.users ?(
+                        group.users.map((user, index) => (
+                            <Text style={styles.buttonText} key={index}>
+                            {user.name}
+                            </Text>
+                        ))
+                        ):(
+                            <Text style={styles.buttonText}> No Users </Text>
+                        )}
+                    </TouchableOpacity>
+                ))}
 
-            <TouchableOpacity style={[styles.button, { backgroundColor: '#E2DDE2' }]}  onPress={() => navigation.navigate('CreateGroup')} >
-                <TabBarIcon name="add" style = {styles.icon}/>
-                <Text style={styles.titleText}></Text>
-            </TouchableOpacity>
-        </View>
-    </CanDoScrollView>
-  );
+                {/*add group button*/}
+                <TouchableOpacity style={[styles.button, { backgroundColor: '#E2DDE2' }]}  onPress={() => navigation.navigate('createGroup')} >
+                    <TabBarIcon name="add" style = {styles.icon}/>
+                    <Text style={styles.titleText}></Text>
+                </TouchableOpacity>
+            </View>
+        </CanDoScrollView>
+    );
 }
 
 const styles = StyleSheet.create({
