+++ app/(tabs)/leaderboard.tsx
@@ -5,104 +5,129 @@ import { ThemedText } from '@/components/ThemedText';
 import CanDoScrollView from '@/components/CanDoScrollView';
 import { Colors } from '@/constants/Colors';
 import { useState, useEffect } from 'react';
-
 import { getApp } from "firebase/app";
 import { collection, addDoc, getFirestore, onSnapshot, query, where } from "firebase/firestore";
 import { app } from "@/app/init";
-// A.K.A. Task List
 
+//database connection
 const db = getFirestore(app)
 
 export default function LeaderboardScreen() {
-  useEffect(() => {
-          const getGroups = collection(db, 'Groups')
+    //state variables
+    const [groups, setGroups] = useState([]);
+    const [users, setUsers] = useState([]);
+    const [group, setGroup] = useState(null);
 
-          const subscribe = onSnapshot(getGroups, (querySnapshot) => {
-              const addGroups = []
-              querySnapshot.forEach((list) => {
-                  addGroups.push({...list.data(), id: list.id})
-              })
-              setGroups(addGroups)
-              console.log(addGroups)
-          });
-          return () => subscribe();
-        }, []);
-  const [groups, setGroups] = useState([]);
-  const data = [
-      {name: "Test1", score: 10},
-      {name: "Test2", score: 50},
-      {name: "Test3", score: 70},
-      {name: "Test4", score: 25},
-      {name: "Test5", score: 15},
-  ]
-  const sortedData = [...data].sort((a, b) => b.score - a.score);
-  const topThree = sortedData.slice(0, 3)
-  const colorScheme = useColorScheme() ?? 'light';
+    //use effect tog et groups from the database
+    useEffect(() => {
+        const getGroups = collection(db, 'Groups')
+        const subscribe = onSnapshot(getGroups, (querySnapshot) => {
+            const addGroups = []
+            querySnapshot.forEach((list) => {
+                addGroups.push({...list.data(), id: list.id})
+            })
+            setGroups(addGroups)
+        });
+        return () => subscribe();
+    }, []);
+
+    //function to handle group selection from the picker
+    const handleGroup = (groupId) => {
+        setGroup(groupId)
+        const chosenGroup = groups.find((group) => group.id === groupId);
+        if(chosenGroup.users){
+            setUsers(chosenGroup.users);
+        }
+    }
+
+    //sort data based on the score
+    const sortedData = [...users].sort((a, b) => b.score - a.score)
+    //get the top three users
+    const topThree = sortedData.slice(0, 3)
+
+    const colorScheme = useColorScheme() ?? 'light';
     const tint = colorScheme === 'dark' ? Colors.dark.tint : Colors.light.tint;
-  return (
-    <CanDoScrollView>
-    <View style={styles.row}>
-                <ThemedText style={styles.settingsName}>Choose Group</ThemedText>
-                    <Picker style={styles.dropdown} dropdownIconColor={tint} itemStyle={styles.dropdownItem}>
-                         {groups.map((group) => (
-                                     <Picker.Item label={group.groupName} value={group.id} />
-                                     ))}
-                    </Picker>
+
+    return (
+        <CanDoScrollView>
+            <View style={styles.labeledInputContainer}>
+                {/*dropdown to choose group*/}
+                <ThemedText style={styles.label}>Choose Group</ThemedText>
+                <Picker selectedValue={group} onValueChange={handleGroup}
+                style={styles.dropdown} dropdownIconColor={tint} itemStyle={styles.dropdownItem}>
+                {groups.map((group) => (
+                    <Picker.Item label={group.groupName} value={group.id} key={group.id}/>
+                ))}
+                </Picker>
             </View>
-      <View style={styles.headerRow}>
-        <View style={styles.leftColumn}>
-           <Text style={styles.headerText}>{topThree[1]?.name}</Text>
-           <Text style={styles.two}>2</Text>
-        </View>
-        <View style={styles.centerColumn}>
-            <Text style={styles.headerText}>{topThree[0]?.name}</Text>
-            <Text style={styles.one}>1</Text>
-        </View>
-        <View style={styles.rightColumn}>
-            <Text style={styles.headerText}>{topThree[2]?.name}</Text>
-            <Text style={styles.three}>3</Text>
-        </View>
-      </View>
-      <FlatList data={sortedData} keyExtractor={(item) => item.name} renderItem={({ item, index }) => (
-          <View style={styles.row}>
-            <View style={styles.column}>
-                <Text style={styles.rowText}>{index + 1}</Text>
+
+            {/*displays the top three*/}
+            <View style={styles.headerRow}>
+                <View style={styles.leftColumn}>
+                    <Text style={styles.headerText}>{topThree[1]?.name}</Text>
+                    <Text style={styles.two}>2</Text>
+                </View>
+                <View style={styles.centerColumn}>
+                    <Text style={styles.headerText}>{topThree[0]?.name}</Text>
+                    <Text style={styles.one}>1</Text>
+                </View>
+                <View style={styles.rightColumn}>
+                    <Text style={styles.headerText}>{topThree[2]?.name}</Text>
+                    <Text style={styles.three}>3</Text>
+                </View>
             </View>
-            <View style={styles.column}>
-                <Text style={styles.rowText}>{item.name}</Text>
-            </View>
-            <View style={styles.column}>
-                <Text style={styles.rowText}>{item.score}</Text>
-            </View>
-          </View>
-          )}
-          />
-    </CanDoScrollView>
-  );
+
+            {/*displays all the users and scores of that group*/}
+            <FlatList data={sortedData} keyExtractor={(item, index) => index.toString()} renderItem={({ item, index }) => (
+                <View style={styles.row}>
+                    <View style={styles.column}>
+                        <Text style={styles.rowText}>{index + 1}</Text>
+                    </View>
+                    <View style={styles.column}>
+                        <Text style={styles.rowText}>{item.name}</Text>
+                    </View>
+                    <View style={styles.column}>
+                        <Text style={styles.rowText}>{item.score}</Text>
+                    </View>
+                </View>
+            )}
+            />
+        </CanDoScrollView>
+    );
 }
 const styles = StyleSheet.create({
      dropdown: {
         color: 'white',
         flexShrink: 1,
         flexGrow: 1,
-        height: 90
-      },
-      dropdownItem: {
+        height: 90,
+     },
+     dropdownItem: {
         color: 'white',
-        fontSize: 20,
+        fontSize: 12,
         height: 90
-      },
-      settingsName: {
+     },
+     groupName: {
         marginRight: 10,
         color: 'white',
-        fontSize: 20
+        fontSize: 12
+     },
+    label: {
+        verticalAlign: 'middle',
+        flex: 1,
+        margin: 14,
+        color: 'white'
       },
+    labeledInputContainer: {
+        flex: 1,
+        flexDirection: 'row'
+    },
     headerRow: {
         flexDirection: 'row',
         paddingVertical: 10,
         paddingHorizontal: 20,
         borderBottomWidth: 1,
-        borderBottomColor: '#ddd',
+        borderBottomColor: 'grey',
         marginBottom: 10,
     },
     leftColumn: {
@@ -166,11 +191,11 @@ const styles = StyleSheet.create({
         paddingVertical: 15,
         paddingHorizontal: 20,
         borderBottomWidth: 1,
-        borderBottomColor: '#ddd',
-        },
+        borderBottomColor: 'grey',
+    },
     rowText: {
         fontSize: 20,
         textAlign: 'center',
         color: 'white',
-        },
-    });
+    },
+});
