+++ app/(tabs)/personalTaskList.tsx
@@ -1,74 +1,98 @@
-import { Image, StyleSheet, View, TouchableOpacity, Alert, Text } from 'react-native';
-
+import { Image, StyleSheet, View, TouchableOpacity, Alert, Text, FlatList } from 'react-native';
+import { collection, addDoc, doc, getDocs, deleteDoc, updateDoc, getFirestore, app, onSnapshot, query, where, whereArrayContains } from "firebase/firestore";
 import ParallaxScrollView from '@/components/ParallaxScrollView';
 import { ThemedText } from '@/components/ThemedText';
 import CanDoScrollView from '@/components/CanDoScrollView';
 import { CheckBox } from 'react-native-elements'
-import { useState } from 'react'
+import { useState, useEffect } from 'react'
+import {createStaticNavigation,useNavigation, useRoute} from '@react-navigation/native';
+
+//database connection
+const db = getFirestore(app)
 
 export default function LeaderboardScreen() {
-  const [checked, setChecked] = useState({
-      aTask1: false,
-      aTask2: false,
-      wTask1: false,
-      wTask2: false,
-      hTask1: false,
-      hTask2: false,
+    const navigation = useNavigation();
+    //state variables
+    const [checked, setChecked] = useState({});
+    const [tasks, setTasks] = useState([])
 
-  });
+    //Function to navigate to task details when a task is clicked
+    const handleTask = (taskId, taskName, taskDescription, taskFrequency, taskPoints) => {
+        //navigates and passes task information as parameters
+        navigation.navigate('taskDetail', {taskId, taskName, taskDescription, taskFrequency, taskPoints});
+    }
 
-  const toggleCheckbox = (task) => {
-    setChecked((prev) => ({
-        ...prev,
-        [task]: !prev[task],
-    }));
-  };
+    //function ot handle checkboxes
+    const handleCheckbox = (task) => {
+        setChecked((prev) => ({
+            ...prev,
+            [task]: !prev[task],
+        }));
+    };
 
-  return (
-    <CanDoScrollView>
-     <View style = {styles.container}>
-        <ThemedText style={styles.title}>All Tasks:</ThemedText>
-        <View style = {styles.titleContainer}>
-            <TouchableOpacity style={styles.taskButton} onPress={() => Alert.alert('Task Button pressed')} >
-                <Text style={styles.taskText}>Task 1</Text>
-                <CheckBox checked={checked.aTask1} onPress = {() => toggleCheckbox('aTask1')} containerStyle = {styles.checkboxContainer}/>
-            </TouchableOpacity>
-        </View>
-        <View style = {styles.titleContainer}>
-            <TouchableOpacity style={styles.taskButton} onPress={() => Alert.alert('Task Button pressed')} >
-                <Text style={styles.taskText}>Task 2</Text>
-                <CheckBox checked={checked.aTask2} onPress = {() => toggleCheckbox('aTask2')} containerStyle = {styles.checkboxContainer}/>
-            </TouchableOpacity>
-        </View>
-        <ThemedText style={styles.title}>Work Tasks:</ThemedText>
-        <View style = {styles.titleContainer}>
-            <TouchableOpacity style={styles.taskButton} onPress={() => Alert.alert('Task Button pressed')} >
-                <Text style={styles.taskText}>Task 1</Text>
-                <CheckBox checked={checked.wTask1} onPress = {() => toggleCheckbox('wTask1')} containerStyle = {styles.checkboxContainer}/>
-            </TouchableOpacity>
-        </View>
-        <View style = {styles.titleContainer}>
-            <TouchableOpacity style={styles.taskButton} onPress={() => Alert.alert('Task Button pressed')} >
-                <Text style={styles.taskText}>Task 2</Text>
-                <CheckBox checked={checked.wTask2} onPress = {() => toggleCheckbox('wTask2')} containerStyle = {styles.checkboxContainer}/>
-            </TouchableOpacity>
-        </View>
-        <ThemedText style={styles.title}>Home Tasks:</ThemedText>
-        <View style = {styles.titleContainer}>
-            <TouchableOpacity style={styles.taskButton} onPress={() => Alert.alert('Task Button pressed')} >
-                <Text style={styles.taskText}>Task 1</Text>
-                    <CheckBox checked={checked.hTask1} onPress = {() => toggleCheckbox('hTask1')} containerStyle = {styles.checkboxContainer}/>
-            </TouchableOpacity>
-        </View>
-        <View style = {styles.titleContainer}>
-            <TouchableOpacity style={styles.taskButton} onPress={() => Alert.alert('Task Button pressed')} >
-                <Text style={styles.taskText}>Task 2</Text>
-                <CheckBox checked={checked.hTask2} onPress = {() => toggleCheckbox('hTask2')} containerStyle = {styles.checkboxContainer}/>
-            </TouchableOpacity>
-        </View>
-     </View>
-    </CanDoScrollView>
-  );
+    //gets the task information from the database
+    useEffect(() => {
+        const getTasks = collection(db, 'Tasks')
+
+        const subscribe = onSnapshot(getTasks, (querySnapshot) => {
+            const addTasks = []
+            querySnapshot.forEach((list) => {
+                addTasks.push({...list.data(), id: list.id})
+            })
+            setTasks(addTasks)
+            console.log(addTasks)
+        });
+        return () => subscribe();
+    }, []);
+
+    const personalTasks = tasks.filter(task => task.groupName === 0)
+    const homeTasks = tasks.filter(task => task.category === 'home')
+    const workTasks = tasks.filter(task => task.category === 'work')
+
+    return (
+        <CanDoScrollView>
+            <View style = {styles.container}>
+                {/*displays all tasks*/}
+                <ThemedText style={styles.title}>All Tasks:</ThemedText>
+                <View style={{ flexDirection: 'row' }}>
+                    <FlatList data={personalTasks} keyExtractor = {(item) => item.id} renderItem={({item}) => (
+                        <TouchableOpacity style={styles.taskButton} onPress={() => handleTask(item.id, item.taskName, item.description, item.frequency, item.points)}>
+                            <Text style={[styles.centerText,{textAlign: 'center'}]}>{item.taskName}</Text>
+                            <View style={{ flex: 5 }} />
+                            <CheckBox checked={checked[item.id] || false} onPress = {() => handleCheckbox(item.id)} containerStyle = {styles.checkboxContainer}/>
+                        </TouchableOpacity>
+                    )}
+                    />
+                </View>
+
+                {/*displays work tasks*/}
+                <ThemedText style={styles.title}>Work Tasks:</ThemedText>
+                <View style={{ flexDirection: 'row' }}>
+                    <FlatList data={workTasks} keyExtractor = {(item) => item.id} renderItem={({item}) => (
+                        <TouchableOpacity style={styles.taskButton} onPress={() => handleTask(item.id, item.taskName, item.description, item.frequency, item.points)}>
+                            <Text style={[styles.centerText,{textAlign: 'center'}]}>{item.taskName}</Text>
+                            <View style={{ flex: 5 }} />
+                            <CheckBox checked={checked[item.id] || false} onPress = {() => handleCheckbox(item.id)} containerStyle = {styles.checkboxContainer}/>
+                        </TouchableOpacity>
+                    )}
+                    />
+                </View>
+
+                {/*displays home tasks*/}
+                <ThemedText style={styles.title}>Home Tasks:</ThemedText>
+                <View style={{ flexDirection: 'row' }}>
+                    <FlatList data={homeTasks} keyExtractor = {(item) => item.id} renderItem={({item}) => (
+                        <TouchableOpacity style={styles.taskButton} onPress={() => handleTask(item.id, item.taskName, item.description, item.frequency, item.points)}>
+                            <Text style={[styles.centerText,{textAlign: 'center'}]}>{item.taskName}</Text>
+                            <View style={{ flex: 5 }} />
+                            <CheckBox checked={checked[item.id] || false} onPress = {() => handleCheckbox(item.id)} containerStyle = {styles.checkboxContainer}/>
+                        </TouchableOpacity>
+                    )}
+                    />
+                </View>
+            </View>
+        </CanDoScrollView>
+    );
 }
 
 const styles = StyleSheet.create({
@@ -83,7 +107,7 @@ const styles = StyleSheet.create({
         color: '#0353A4',
 
     },
-    taskText: {
+    centerText: {
         color: 'white',
         fontSize: 15,
     },
